requirements.txt
groq
python-dotenv
SpeechRecognition
pydub
elevenlabs
gTTS
